Move Emails models from super_archieves to core
EmailAddress and EmailAddressValidation moved from
super_archieves.models to accounts.models.
