+++ colab/accounts/templates/emails/email_verification.txt
