+++ colab/accounts/utils/email.py
@@ -13,7 +13,7 @@ def colab_send_email(subject, message, to):
 def send_verification_email(to, user, validation_key, verification_url):
     subject = _('Please verify your email ') + u'{}'.format(to)
     msg_tmpl = \
-        loader.get_template('superarchives/emails/email_verification.txt')
+        loader.get_template('emails/email_verification.txt')
     message = msg_tmpl.render(Context({'to': to, 'user': user,
                                        'verification_url': verification_url
                                        }))
