+++ colab/super_archives/models.py
@@ -2,9 +2,6 @@
 
 import urllib
 
-from uuid import uuid4
-from hashlib import md5
-
 from django.db import models
 from django.conf import settings
 from django.utils import timezone
@@ -17,67 +14,10 @@ from taggit.managers import TaggableManager
 from hitcounter.models import HitCounterModelMixin
 
 from .managers import NotSpamManager, MostVotedManager, HighestScore
-from .utils import blocks, email
+from .utils import blocks
 from .utils.etiquetador import etiquetador
 from colab.accounts.utils import mailman
-
-
-def get_validation_key():
-    return uuid4().hex
-
-
-class EmailAddressValidation(models.Model):
-    address = models.EmailField(unique=True)
-    user = models.ForeignKey(settings.AUTH_USER_MODEL, null=True,
-                             related_name='emails_not_validated')
-    validation_key = models.CharField(max_length=32, null=True,
-                                      default=get_validation_key)
-    created = models.DateTimeField(auto_now_add=True)
-
-    class Meta:
-        unique_together = ('user', 'address')
-
-    @classmethod
-    def create(cls, address, user):
-        email_address_validation = cls.objects.create(address=address,
-                                                      user=user)
-        return email_address_validation
-
-    @classmethod
-    def verify_email(cls, email_address_validation, verification_url):
-        return email.send_verification_email(
-            email_address_validation.address,
-            email_address_validation.user,
-            email_address_validation.validation_key,
-            verification_url
-            )
-
-
-class EmailAddress(models.Model):
-    user = models.ForeignKey(settings.AUTH_USER_MODEL, null=True,
-                             related_name='emails', on_delete=models.SET_NULL)
-    address = models.EmailField(unique=True)
-    real_name = models.CharField(max_length=64, blank=True, db_index=True)
-    md5 = models.CharField(max_length=32, null=True)
-
-    class Meta:
-        ordering = ('id', )
-
-    def save(self, *args, **kwargs):
-        self.md5 = md5(self.address).hexdigest()
-        super(EmailAddress, self).save(*args, **kwargs)
-
-    def get_full_name(self):
-        if self.user and self.user.get_full_name():
-            return self.user.get_full_name()
-        else:
-            return self.real_name
-
-    def get_full_name_or_anonymous(self):
-        return self.get_full_name() or _('Anonymous')
-
-    def __unicode__(self):
-        return '"%s" <%s>' % (self.get_full_name(), self.address)
+from colab.accounts.models import EmailAddress
 
 
 class MailingList(models.Model):
