+++ colab/super_archives/signals.py
@@ -3,7 +3,7 @@ from django.db.models.signals import post_save
 from django.dispatch import receiver
 from django.conf import settings
 
-from .models import EmailAddress
+from colab.accounts.models import EmailAddress
 
 
 @receiver(post_save, sender=settings.AUTH_USER_MODEL)
