+++ colab/super_archives/views.py
@@ -20,10 +20,9 @@ from django.contrib.auth.decorators import login_required
 from django.shortcuts import render, redirect, get_object_or_404
 
 from colab.accounts.utils import mailman
-from colab.accounts.models import User
+from colab.accounts.models import (User, EmailAddress, EmailAddressValidation)
 from .utils.email import send_verification_email
-from .models import (MailingList, Thread, EmailAddress,
-                     EmailAddressValidation, Message)
+from .models import (MailingList, Thread, Message)
 
 
 class ThreadView(View):
